chore: 회원 컴포넌트명 변경 및 컴포넌트 분리
